Define basic auto build system for the project
In this commit, I prepare a simple self-contained build environment,
using docker (podman) and a Makefile to automatically build the project
in a supportive build environment. For now, the user will still have to
install a few support libraries if he wants to actually run the
executables, but without the custom hardware there isn't much he can do
anyway, so that's not too important now.