Define weave and convert Makefile to semi-literate form
In this commit, I update the Makefile to use semi-literate form. This allows me to capture useful documentation within the source itself. I can
generalize weave to apply to other languages, if Doxygen is not the best place for that documentation. However, I think for now, weave will remain
dedicate to the Makfile.